Remove duplicate interaction componenets and MovieCard Component rename to PosterCard for more clarity
