+++ media-wrangler-front-end/src/components/MovieInteractions/StarRatingButton.jsx
