+++ media-wrangler-front-end/src/components/MoviePosterCard/PosterCard.css
